Created repositories, services and controllers

+++ src/main/java/mk/ukim/finki/easyfood/repository/DeliveryFirmRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DeliveryFirmRepository extends JpaRepository<DeliveryFirm, Integer> {
+public interface DeliveryFirmRepository extends JpaRepository<DeliveryFirm, Long> {
 }
